Implement draft/publish workflow with version history integration and confirmation dialog; refactor SolidFormView to utilize new workflow hooks
